requirements.txt
appnope==0.1.3
asttokens==2.4.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
html2text==2020.1.16
idna==3.6
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
Wikipedia-API==0.6.0
wikirate4py @ git+https://github.com/wikirate/wikirate4py.git@a8d8821ae8b1f79be2dc62422a7b5c8bc751d581
